cases: fix cases without a reason not being posted
The fact that cases *with* a reason were being posted was actually an
unintended side effect of createCaseNote also posting the case after
creating the note. Had the issue with the cases not being posted not
existed, new cases with a reason would've been posted twice. This commit
changes this so createCaseNote doesn't post the case when used from
createCase, and fixes the actual case posting at the end of createCase.