Tweaks to Roles plugin before merging
Separate role adding/removing to two separate commands for clearer
help pages and to remove the conflict with the self-assignable roles
plugin.

Require the assignable_roles option to always be an array (even if an
empty one).

Allow role assignments to self.

Log role additions/removals via these commands with the right
moderator name (instead of relying on the auto-log from the event).