Persist: remove voice mute persist support
Since voice mutes cannot be re-applied unless the member is in voice,
we can't re-apply the voice mute automatically on join and would have
to do it later when they first join voice, which could get messy if
they were not supposed to be voice muted anymore after all (if they
e.g. don't use voice for a long time).