refactor: change how DEBUG behaves in prod
The env var now only affects the bot container, not the API.
This is because we don't currently need this extra debug information
for the API.

For the bot, with DEBUG set, the bot container will now start and do
nothing. To start the bot in this container in debug mode, open a
shell into the container and run the start-bot-prod-debug script
manually. This change was made to allow more control over what
profiling scripts and other things are run in the debug environment.