fix: standalone migrations
Recent fix for dropping/recreating PKs in environments with
sql_require_primary_key=1 also broke the migrations for environments
where the user doesn't have permissions to change the value of the
setting. The standalone setup is such an environment.

This commit changes those migrations so that the PKs are
dropped/recreated as one ALTER TABLE query, no longer requiring
changing the sql_require_primary_key setting.